iommupt: Add the basic structure of the iommu implementation
The iommu implementation is a single version of the iommu domain
operations, iova_to_phys, map, unmap, read_and_clear_dirty and
flushing. It is intended to be a near drop in replacement for existing
iopt users.

By using the Generic Page Table mechanism it is a single algorithmic
implementation that operates all the different page table formats with
consistent characteristics.

Implement the basic starting point: alloc(), get_info() and deinit().

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>