iommupt: Add map_range op
Implement a self-segmenting algorithm for map_range. This can handle any
valid input VA/length and will automatically break it up into
appropriately sized table entries using a recursive descent algorithm.

The appropriate page size is computed each step using some bitwise
calculations.

map is slightly complicated because it has to handle a number of special
edge cases:
 - Overmapping a previously shared table with an OA - requries validating
   and discarding the possibly empty tables
 - Doing the above across an entire to-be-created contiguous entry
 - Installing a new table concurrently with another thread
 - Racing table installation with CPU cache flushing
 - Expanding the table by adding more top levels on the fly

Managing the table installation race is done using a flag in the
folio. When the shared table entry is possibly unflushed the flag will be
set. This works for all pagetable formats but is less efficient than the
io-pgtable-arm-lpae approach of using a SW table bit. It may be
interesting to provide the latter as an option.

Table expansion is a unique feature of AMDv1, this version is quite
similar except we handle racing concurrent lockless map. The table top
pointer and starting level are encoded in a single uintptr_t which ensures
we can READ_ONCE() without tearing. Any op will do the READ_ONCE() and use
that fixed point as its starting point. Concurrent expansion is handled
with a table global spinlock.

When inserting a new table entry map checks that the portion of the table
is empty. This includes removing an empty interior tables. The approach
here is atomic per entry. Either the new entry is written, or no change is
made to the table. This is done by keeping a list of interior tables to
free and only progressing once the entire space is checked to be empty.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>