cover-letter: Consolidate iommu page table implementations
Currently each of the iommu page table formats duplicates all of the logic
to maintain the page table and perform map/unmap/etc operations. There are
several different versions of the algorithms between all the different
formats. The io-pgtable system provides an interface to help isolate the
page table code from the iommu driver, but doesn't provide tools to
implement the common algorithms.

This makes it very hard to improve the state of the pagetable code under
the iommu domains as any proposed improvement needs to alter a large
number of different driver code paths. Combined with a lack of software
based testing this makes improvement in this area very hard.

iommufd wants several new page table operations:
 - More efficient map/unmap operations, using iommufd's batching approach
 - unmap that returns the physical addresses into a batch as it progresses
 - cut that allows splitting areas so large pages can have holes
   poked in them dynamically
 - More agressive freeing of table memory to avoid waste
 - Fragmenting large pages so that dirty tracking can run efficiently
 - Reassembling large pages so that VMs can run at full IO performance
   in error flows
 - guestmemfd hitless shared/private transitions

In addition there are possibilities like directly mapping a bvec, or
sg_list in more efficient ways, and perhaps even optimizations for the GPU
drivers using the io-pgtable code as well.

Together these are algorithmically complex enough to be a very significant
task to go and implement in all the page table formats we support. Just
the "server" focused drivers use almost all the formats (ARMv8 S1&S2 / x86
PAE / AMDv1 / VT-D SS / RISCV)

Instead of doing the duplicated work, this series takes the first step to
consolidate the algorithms into one places. In spirit it is similar to the
work Christoph did a few years back to pull the redundant get_user_pages()
implementations out of the arch code into core MM. This unlocked a great
deal of improvement in that space in the following years. I would like to
see the same benefit in iommu as well.

The approach is split into three deliberate layers:
 - The truely generic page table components. These are very application
   neutral and could conceivably by used in the MM or KVM if there was
   a reason. A DRM driver may also be interested in this layer as it
   could be more efficient than working through the iommu focused ops.

 - The per format functions. These are a set of small inline functions
   that abstract the details on how the page table is layed out in
   memory and what bits do what things. Like the MM these functions
   share the same name so the same code can be compiled against
   different formats by including the appropriate format header.

 - An iommu implementation. This is intended to create ops that can
   take over from the iommu_domain ops. There is a single set of C
   routines that compile against all the formats generically.

On top of this are two kunit tests, one that directly exercises the iommu
implementation across all the different formats. The second kunit does an
A/B comparison between the iommupt and the io-pgtable implementation to
ensure things are identical.

Sort of like MM, this uses multi-compilation where the common code
includes format specific headers that implement the same C API. Unlike the
MM we need to build multiple page table formats into the same kernel, so
each combinaton of format/parameters/iommu implementation is compiled in a
single compilation unit and into a module. This results in compiling the
same C code multiple times in a single kernel build, using different
combinations of header files.

The approach is designed to be able to provide both mm-like fully inlined
performance, or as typical for iommu, recursive non-inlined smaller .text
version. As the implementation is now shared it will be worthwhile to do
some performance work and fine tune this as appropriate.

I've CC'd a few people from outside iommu that may have some interest in
the generic part of this, or ideas how to better build the abstraction and
helpers.

For this RFC I've provided draft formats for nearly everything (S390 and
RISCV are notably not included). The formats all pass the compare test and
thus, to a significant degree, produce the same memory layouts for the
radix tree. The primary purpose of this breadth is to prove the common API
is suitable for the job. Completing these to be fully usable in their
respective drivers is still to be done.

I'm expecting to show maybe another RFC round with all the formats and
pivot to a more focused series, likely just for AMD, that brings the
minimum necessary. From there we can work in parallel to add the new
iommufd features and convert more of the drievrs. From an iommufd
perspective I would like the "server" drivers (AMD / SMMUv3 / VT-D) to be
converted as a minimum.

This general concept was brough up and discussed a few times during LPC
last year and I have a formal session on the schedule for this series in
LPC Vienna.

There are many additional support patches required to run the kunits,
everything is on github:

https://github.com/jgunthorpe/linux/commits/iommu_pt

FIXME:
  - Improve the two kunit tests
  - Implement additional new iommufd ops
  - Implement the debugfs with the RCU safety
  - Do a performance study vs the io-pgtable versions
  - Implement the flush callbacks, iommu core hookups, etc
  - Look at possible bvec and sg optimizations
  - Link it up to the iommu drivers and test it in HW as an iommu
    implementation

Cc: Joao Martins <joao.m.martins@oracle.com>
Cc: Alejandro Jimenez <alejandro.j.jimenez@oracle.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Pasha Tatashin <pasha.tatashin@soleen.com>
Cc: Peter Xu <peterx@redhat.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Sean Christopherson <seanjc@google.com>
Cc: Ryan Roberts <ryan.roberts@arm.com>
Cc: iommu@lists.linux.dev
Cc: kvm@vger.kernel.org
Cc: linux-mm@kvack.org
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>