iommupt: Add cut_mapping op
The cut operation allows breaking large pages into smaller pages with
specific purpose of allowing unmapping of a portion of the previously
large page. Specifically it ensures that no large page crosses the cut
point, and thus, we can successfully unmap starting/ending on the cut
point.

This is the operation that VFIO type 1 v1.0 implicitly supported and some
iommu drivers provided internal to their unmap operation.

Implement the cut operation to be hitless, changes to the page table
during cutting must cause zero disruption to any ongoing DMA. This is the
expectation of the VFIO type 1 uAPI. Hitless requires HW support, it is
incompatible with HW requiring break-before-make.

Having it separate from unmap makes it much easier to handle failure
cases. Since cut is fully hitless, even in failure cases, a caller can
simply do nothing if cut fails. Cut during unmap requries dealing with the
potentially nasty case where some of the IOVA range has been unmapped but
some cannot be unmapped.

The operation is generalized into a form that iommufd can use for it's
existing area split operation. By placing cuts on the boundaries of the
split we can subdivide an area and maintain all existing semantics even
with large pages in the page table.

Cut is optimal and generates a page table that is equivalent to calling
map twice, that is the two halves will still use maximal page sizes.

FIXME: requires deeper kunit tests

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>