iommupt: Add the AMD IOMMU v1 page table format
AMD IOMMU v1 is unique in supporting contiguous pages with a variable
size and it can decode the full 64 bit VA space.

The general design is quite similar to the x86 PAE format, except with an
additional level and quite different PTE encoding.

This format is the only one that uses the PT_FEAT_DYNAMIC_TOP feature in
the existing code.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>