iommupt: Add a mock pagetable format for iommufd selftest to use
This is a minor tweak to the amdv1 format which is quite suitable for
the iommufd self test. It is adjusted to have a 2K page granule to continue
allow testing the corner case of an IOMMU page table that is smaller than
the host (ie a 4k table on 64k ARM or something).

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>