cover-letter: iommu: Further abstract iommu-pages
iommu-pages is a small abstraction for allocating page table pages that
iommu drivers use. It has a few special properties that distinguish it
from the other allocators in the kernel:

 - Allocations are always power of two, and always physically aligned to
   their size
 - Allocations can be threaded on a list, and all allocations on the list
   can be freed. This must be done without memory allocations
 - Allocations are accounted for in the secondary page table counters
 - Allocations can sometimes be less than a full CPU page, 1/4 and 1/16 are
   some common sub page allocation sizes.
 - Allocations can sometimes be multiple CPU pages

Make the API tighter and leak fewer internal details to the
callers. Particularly this series aims to remove all struct page usage
related to iommu-pages memory from all drivers, this is things such as:

 struct page
 virt_to_page()
 page_to_virt()
 page_to_pfn()
 pfn_to_page()
 dma_map_page()
 page_address()
 page->lru

Once drivers no longer use struct page convert iommu-pages to folios and
use a private memory descriptor. This slightly progresses Matthew's
memdesc project and clears the way to using more space in the struct page
for iommu-pages features in future.

Improve the API to work directly on sizes instead of order, the drivers
generally have HW specs and code paths that already in specific sizes.
Pass those sizes down into the allocator to remove some boiler plate in
drivers and be more prepared for if there is some future sub page
allocator work.

Cc: Matthew Wilcox <willy@infradead.org>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>