vfio/type1: Handle 64-bit address space overflow on unmap
DMA mappings that extend up to the end of the 64-bit IOVA space can be
created, but cannot be found while traversing the dma_list, causing
failures in VFIO_IOMMU_UNMAP_DMA for such mappings. This is due to integer
overflow affecting the calculation of overlapping ranges.

Ensure that overflow is correctly handled in these edge cases, so such
ranges can be unmapped.

Signed-off-by: Alejandro Jimenez <alejandro.j.jimenez@oracle.com>

---

Reproducer:
https://gist.github.com/aljimenezb/f3338c9c2eda9b0a7bf5f76b40354db8

I am aware the minimal addition of " - 1" in a few selected places can be
used to mitigate this issue. However that relies on the enforcement by
other sections of the code that the ranges will not actually wrap around,
and just merely extend to the end of the address space. The implemented
approach is more robust and explicit.